data/requirements.txt
requests>=2.31.0
pandas>=2.1.0
sqlite3
python-dotenv>=1.0.0
schedule>=1.2.0


